Move the send path tests into their own file
Everything between handing a chunk to AceComm and knowing what happened to
it: the ACK timer and its liveness tag, retries, NACK and its backoff, the
pacing levers, the FinishSending diagnostics, and the offline paths.

Two section banners are dropped rather than moved. Both had drifted away from
the describe they named and both said what the describe title already says.
